apps/research-engine/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
playwright==1.47.0
httpx>=0.28.1
google-generativeai==0.8.0
sqlalchemy==2.0.35
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.9.0
